Enum tauri_runtime::RunEvent

#[non_exhaustive]
pub enum RunEvent<T: UserEvent> {
    Exit,
    ExitRequested {
        code: Option<i32>,
        tx: Sender<ExitRequestedEventAction>,
    },
    WindowEvent {
        label: String,
        event: WindowEvent,
    },
    WebviewEvent {
        label: String,
        event: WebviewEvent,
    },
    Ready,
    Resumed,
    MainEventsCleared,
    Opened {
        urls: Vec<Url>,
    },
    Reopen {
        has_visible_windows: bool,
    },
    UserEvent(T),
}

Event triggered on the event loop run.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Exit

Event loop is exiting.

ExitRequested

Event loop is about to exit

Fields

code: Option<i32>

The exit code.

tx: Sender<ExitRequestedEventAction>

WindowEvent

An event associated with a window.

Fields

label: String

The window label.

event: WindowEvent

The detailed event.

WebviewEvent

An event associated with a webview.

Fields

label: String

The webview label.

event: WebviewEvent

The detailed event.

Ready

Application ready.

Resumed

Sent if the event loop is being resumed.

MainEventsCleared

Emitted when all of the event loop’s input events have been processed and redraw processing is about to begin.

This event is useful as a place to put your code that should be run after all state-changing events have been handled and you want to do stuff (updating state, performing calculations, etc) that happens as the “main body” of your event loop.

Opened

Emitted when the user wants to open the specified resource with the app.

Fields

urls: Vec<Url>

Reopen

Emitted when the NSApplicationDelegate’s applicationShouldHandleReopen gets called

Fields

has_visible_windows: bool

Indicates whether the NSApplication object found any visible windows in your application.

UserEvent(T)

A custom event defined by the user.