Module tauri_runtime::menu

Create window and system tray menus.

Structs

AboutMetadata

Application metadata for the MenuItem::About action.

CustomMenuItem

A custom menu item.

Menu

A window menu.

Submenu

SystemTrayMenu

A system tray menu.

SystemTraySubmenu

Enums

MenuEntry

An entry on the system tray menu.

MenuItem

A menu item, bound to a pre-defined action or Custom emit an event. Note that status bar only supports Custom menu item variants. And on the menu bar, some platforms might not support some of the variants. Unsupported variant will be no-op on such platform.

MenuUpdate

SystemTrayMenuEntry

An entry on the system tray menu.

SystemTrayMenuItem

System tray menu item.

Traits

TrayHandle

Type Definitions

MenuHash

MenuId

MenuIdRef