Struct tauri_runtime::menu::Menu

#[non_exhaustive]
pub struct Menu {
    pub items: Vec<MenuEntry>,
}

A window menu.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

items: Vec<MenuEntry>
Implementations

impl Menu

pub fn new() -> Self

Creates a new window menu.

pub fn os_default(app_name: &str) -> Self

Creates a menu filled with default menu items and submenus.

Platform-specific:

Windows:

File

CloseWindow

Quit

Edit

Cut

Copy

Paste

Window

Minimize

CloseWindow

Linux:

File

CloseWindow

Quit

Window

Minimize

CloseWindow

macOS:

App

About

Separator

Services

Separator

Hide

HideOthers

ShowAll

Separator

Quit

File

CloseWindow

Edit

Undo

Redo

Separator

Cut

Copy

Paste

SelectAll

View

EnterFullScreen

Window

Minimize

Zoom

Separator

CloseWindow

pub fn with_items<I: IntoIterator<Item = MenuEntry>>(items: I) -> Self

Creates a new window menu with the given items.

Examples

Menu::with_items([
  MenuItem::SelectAll.into(),
  #[cfg(target_os = "macos")]
  MenuItem::Redo.into(),
  CustomMenuItem::new("toggle", "Toggle visibility").into(),
  Submenu::new("View", Menu::new()).into(),
]);

pub fn add_item(self, item: CustomMenuItem) -> Self

Adds the custom menu item to the menu.

pub fn add_native_item(self, item: MenuItem) -> Self

Adds a native item to the menu.

pub fn add_submenu(self, submenu: Submenu) -> Self

Adds an entry with submenu.