Trait tauri_runtime::Dispatch

pub trait Dispatch<T: UserEvent>: Debug + Clone + Send + Sync + Sized + 'static {
    type Runtime: Runtime<T>;
    type WindowBuilder: WindowBuilder;

    fn run_on_main_thread<F: FnOnce() + Send + 'static>(
        &self, 
        f: F
    ) -> Result<()>;
    fn on_window_event<F: Fn(&WindowEvent) + Send + 'static>(
        &self, 
        f: F
    ) -> Uuid;
    fn on_menu_event<F: Fn(&MenuEvent) + Send + 'static>(&self, f: F) -> Uuid;
    fn open_devtools(&self);
    fn close_devtools(&self);
    fn is_devtools_open(&self) -> Result<bool>;
    fn scale_factor(&self) -> Result<f64>;
    fn inner_position(&self) -> Result<PhysicalPosition<i32>>;
    fn outer_position(&self) -> Result<PhysicalPosition<i32>>;
    fn inner_size(&self) -> Result<PhysicalSize<u32>>;
    fn outer_size(&self) -> Result<PhysicalSize<u32>>;
    fn is_fullscreen(&self) -> Result<bool>;
    fn is_maximized(&self) -> Result<bool>;
    fn is_decorated(&self) -> Result<bool>;
    fn is_resizable(&self) -> Result<bool>;
    fn is_visible(&self) -> Result<bool>;
    fn is_menu_visible(&self) -> Result<bool>;
    fn current_monitor(&self) -> Result<Option<Monitor>>;
    fn primary_monitor(&self) -> Result<Option<Monitor>>;
    fn available_monitors(&self) -> Result<Vec<Monitor>>;
    fn gtk_window(&self) -> Result<ApplicationWindow>;
    fn raw_window_handle(&self) -> Result<RawWindowHandle>;
    fn theme(&self) -> Result<Theme>;
    fn center(&self) -> Result<()>;
    fn print(&self) -> Result<()>;
    fn request_user_attention(
        &self, 
        request_type: Option<UserAttentionType>
    ) -> Result<()>;
    fn create_window(
        &mut self, 
        pending: PendingWindow<T, Self::Runtime>
    ) -> Result<DetachedWindow<T, Self::Runtime>>;
    fn set_resizable(&self, resizable: bool) -> Result<()>;
    fn set_title<S: Into<String>>(&self, title: S) -> Result<()>;
    fn maximize(&self) -> Result<()>;
    fn unmaximize(&self) -> Result<()>;
    fn minimize(&self) -> Result<()>;
    fn unminimize(&self) -> Result<()>;
    fn show_menu(&self) -> Result<()>;
    fn hide_menu(&self) -> Result<()>;
    fn show(&self) -> Result<()>;
    fn hide(&self) -> Result<()>;
    fn close(&self) -> Result<()>;
    fn set_decorations(&self, decorations: bool) -> Result<()>;
    fn set_always_on_top(&self, always_on_top: bool) -> Result<()>;
    fn set_size(&self, size: Size) -> Result<()>;
    fn set_min_size(&self, size: Option<Size>) -> Result<()>;
    fn set_max_size(&self, size: Option<Size>) -> Result<()>;
    fn set_position(&self, position: Position) -> Result<()>;
    fn set_fullscreen(&self, fullscreen: bool) -> Result<()>;
    fn set_focus(&self) -> Result<()>;
    fn set_icon(&self, icon: Icon) -> Result<()>;
    fn set_skip_taskbar(&self, skip: bool) -> Result<()>;
    fn set_cursor_grab(&self, grab: bool) -> Result<()>;
    fn set_cursor_visible(&self, visible: bool) -> Result<()>;
    fn set_cursor_icon(&self, icon: CursorIcon) -> Result<()>;
    fn set_cursor_position<Pos: Into<Position>>(
        &self, 
        position: Pos
    ) -> Result<()>;
    fn start_dragging(&self) -> Result<()>;
    fn eval_script<S: Into<String>>(&self, script: S) -> Result<()>;
    fn update_menu_item(&self, id: u16, update: MenuUpdate) -> Result<()>;

}

Webview dispatcher. A thread-safe handle to the webview API.

Required Associated Types

type Runtime: Runtime<T>

The runtime this Dispatch runs under.

type WindowBuilder: WindowBuilder

The window builder type.

Required Methods

fn run_on_main_thread<F: FnOnce() + Send + 'static>(&self, f: F) -> Result<()>

Run a task on the main thread.

fn on_window_event<F: Fn(&WindowEvent) + Send + 'static>(&self, f: F) -> Uuid

Registers a window event handler.

fn on_menu_event<F: Fn(&MenuEvent) + Send + 'static>(&self, f: F) -> Uuid

Registers a window event handler.

fn open_devtools(&self)

Open the web inspector which is usually called devtools.

fn close_devtools(&self)

Close the web inspector which is usually called devtools.

fn is_devtools_open(&self) -> Result<bool>

Gets the devtools window’s current open state.

fn scale_factor(&self) -> Result<f64>

Returns the scale factor that can be used to map logical pixels to physical pixels, and vice versa.

fn inner_position(&self) -> Result<PhysicalPosition<i32>>

Returns the position of the top-left hand corner of the window’s client area relative to the top-left hand corner of the desktop.

fn outer_position(&self) -> Result<PhysicalPosition<i32>>

Returns the position of the top-left hand corner of the window relative to the top-left hand corner of the desktop.

fn inner_size(&self) -> Result<PhysicalSize<u32>>

Returns the physical size of the window’s client area.

The client area is the content of the window, excluding the title bar and borders.

fn outer_size(&self) -> Result<PhysicalSize<u32>>

Returns the physical size of the entire window.

These dimensions include the title bar and borders. If you don’t want that (and you usually don’t), use inner_size instead.

fn is_fullscreen(&self) -> Result<bool>

Gets the window’s current fullscreen state.

fn is_maximized(&self) -> Result<bool>

Gets the window’s current maximized state.

fn is_decorated(&self) -> Result<bool>

Gets the window’s current decoration state.

fn is_resizable(&self) -> Result<bool>

Gets the window’s current resizable state.

fn is_visible(&self) -> Result<bool>

Gets the window’s current visibility state.

fn is_menu_visible(&self) -> Result<bool>

Gets the window menu current visibility state.

fn current_monitor(&self) -> Result<Option<Monitor>>

Returns the monitor on which the window currently resides.

Returns None if current monitor can’t be detected.

fn primary_monitor(&self) -> Result<Option<Monitor>>

Returns the primary monitor of the system.

Returns None if it can’t identify any monitor as a primary one.

fn available_monitors(&self) -> Result<Vec<Monitor>>

Returns the list of all the monitors available on the system.

fn gtk_window(&self) -> Result<ApplicationWindow>

Returns the ApplicationWindow from gtk crate that is used by this window.

fn raw_window_handle(&self) -> Result<RawWindowHandle>

fn theme(&self) -> Result<Theme>

Returns the current window theme.

fn center(&self) -> Result<()>

Centers the window.

fn print(&self) -> Result<()>

Opens the dialog to prints the contents of the webview.

fn request_user_attention(
&self,
request_type: Option<UserAttentionType>
) -> Result<()>

Requests user attention to the window.

Providing None will unset the request for user attention.

fn create_window(
&mut self,
pending: PendingWindow<T, Self::Runtime>
) -> Result<DetachedWindow<T, Self::Runtime>>

Create a new webview window.

fn set_resizable(&self, resizable: bool) -> Result<()>

Updates the window resizable flag.

fn set_title<S: Into<String>>(&self, title: S) -> Result<()>

Updates the window title.

fn maximize(&self) -> Result<()>

Maximizes the window.

fn unmaximize(&self) -> Result<()>

Unmaximizes the window.

fn minimize(&self) -> Result<()>

Minimizes the window.

fn unminimize(&self) -> Result<()>

Unminimizes the window.

fn show_menu(&self) -> Result<()>

Shows the window menu.

fn hide_menu(&self) -> Result<()>

Hides the window menu.

fn show(&self) -> Result<()>

Shows the window.

fn hide(&self) -> Result<()>

Hides the window.

fn close(&self) -> Result<()>

Closes the window.

fn set_decorations(&self, decorations: bool) -> Result<()>

Updates the hasDecorations flag.

fn set_always_on_top(&self, always_on_top: bool) -> Result<()>

Updates the window alwaysOnTop flag.

fn set_size(&self, size: Size) -> Result<()>

Resizes the window.

fn set_min_size(&self, size: Option<Size>) -> Result<()>

Updates the window min size.

fn set_max_size(&self, size: Option<Size>) -> Result<()>

Updates the window max size.

fn set_position(&self, position: Position) -> Result<()>

Updates the window position.

fn set_fullscreen(&self, fullscreen: bool) -> Result<()>

Updates the window fullscreen state.

fn set_focus(&self) -> Result<()>

Bring the window to front and focus.

fn set_icon(&self, icon: Icon) -> Result<()>

Updates the window icon.

fn set_skip_taskbar(&self, skip: bool) -> Result<()>

Whether to show the window icon in the task bar or not.

fn set_cursor_grab(&self, grab: bool) -> Result<()>

Grabs the cursor, preventing it from leaving the window.

There’s no guarantee that the cursor will be hidden. You should hide it by yourself if you want so.

fn set_cursor_visible(&self, visible: bool) -> Result<()>

Modifies the cursor’s visibility.

If false, this will hide the cursor. If true, this will show the cursor.

fn set_cursor_icon(&self, icon: CursorIcon) -> Result<()>

fn set_cursor_position<Pos: Into<Position>>(&self, position: Pos) -> Result<()>

Changes the position of the cursor in window coordinates.

fn start_dragging(&self) -> Result<()>

Starts dragging the window.

fn eval_script<S: Into<String>>(&self, script: S) -> Result<()>

Executes javascript on the window this Dispatch represents.

fn update_menu_item(&self, id: u16, update: MenuUpdate) -> Result<()>

Applies the specified update to the menu item associated with the given id.