Trait tauri_runtime::Runtime

pub trait Runtime<T: UserEvent>: Debug + Sized + 'static {
    type Dispatcher: Dispatch<T, Runtime = Self>;
    type Handle: RuntimeHandle<T, Runtime = Self>;
    type GlobalShortcutManager: GlobalShortcutManager;
    type ClipboardManager: ClipboardManager;
    type EventLoopProxy: EventLoopProxy<T>;

    fn new() -> Result<Self>;
    fn new_any_thread() -> Result<Self>;
    fn create_proxy(&self) -> Self::EventLoopProxy;
    fn handle(&self) -> Self::Handle;
    fn global_shortcut_manager(&self) -> Self::GlobalShortcutManager;
    fn clipboard_manager(&self) -> Self::ClipboardManager;
    fn create_window(
        &self, 
        pending: PendingWindow<T, Self>
    ) -> Result<DetachedWindow<T, Self>>;
    fn run_iteration<F: Fn(RunEvent<T>) + 'static>(
        &mut self, 
        callback: F
    ) -> RunIteration;
    fn run<F: FnMut(RunEvent<T>) + 'static>(self, callback: F);
}

The webview runtime interface.

Required Associated Types

type Dispatcher: Dispatch<T, Runtime = Self>

The message dispatcher.

type Handle: RuntimeHandle<T, Runtime = Self>

The runtime handle type.

type GlobalShortcutManager: GlobalShortcutManager

The global shortcut manager type.

type ClipboardManager: ClipboardManager

The clipboard manager type.

type EventLoopProxy: EventLoopProxy<T>

The proxy type.

Required Methods

fn new() -> Result<Self>

Creates a new webview runtime. Must be used on the main thread.

fn new_any_thread() -> Result<Self>

Creates a new webview runtime on any thread.

fn create_proxy(&self) -> Self::EventLoopProxy

Creates an EventLoopProxy that can be used to dispatch user events to the main event loop.

fn handle(&self) -> Self::Handle

Gets a runtime handle.

fn global_shortcut_manager(&self) -> Self::GlobalShortcutManager

Gets the global shortcut manager.

fn clipboard_manager(&self) -> Self::ClipboardManager

Gets the clipboard manager.

fn create_window(
&self,
pending: PendingWindow<T, Self>
) -> Result<DetachedWindow<T, Self>>

Create a new webview window.

fn run_iteration<F: Fn(RunEvent<T>) + 'static>(
&mut self,
callback: F
) -> RunIteration

Runs the one step of the webview runtime event loop and returns control flow to the caller.

fn run<F: FnMut(RunEvent<T>) + 'static>(self, callback: F)

Run the webview runtime.