Trait tauri_runtime::RuntimeHandle

pub trait RuntimeHandle<T: UserEvent>:
    Debug
    + Clone
    + Send
    + Sync
    + Sized
    + 'static {
    type Runtime: Runtime<T, Handle = Self>;

    // Required methods
    fn create_proxy(&self) -> <Self::Runtime as Runtime<T>>::EventLoopProxy;
    fn set_activation_policy(
        &self,
        activation_policy: ActivationPolicy,
    ) -> Result<()>;
    fn request_exit(&self, code: i32) -> Result<()>;
    fn create_window<F: Fn(RawWindow<'_>) + Send + 'static>(
        &self,
        pending: PendingWindow<T, Self::Runtime>,
        before_window_creation: Option<F>,
    ) -> Result<DetachedWindow<T, Self::Runtime>>;
    fn create_webview(
        &self,
        window_id: WindowId,
        pending: PendingWebview<T, Self::Runtime>,
    ) -> Result<DetachedWebview<T, Self::Runtime>>;
    fn run_on_main_thread<F: FnOnce() + Send + 'static>(
        &self,
        f: F,
    ) -> Result<()>;
    fn display_handle(&self) -> Result<DisplayHandle<'_>, HandleError>;
    fn primary_monitor(&self) -> Option<Monitor>;
    fn monitor_from_point(&self, x: f64, y: f64) -> Option<Monitor>;
    fn available_monitors(&self) -> Vec<Monitor>;
    fn cursor_position(&self) -> Result<PhysicalPosition<f64>>;
    fn set_theme(&self, theme: Option<Theme>);
    fn show(&self) -> Result<()>;
    fn hide(&self) -> Result<()>;

}

A Send handle to the runtime.

Required Associated Types

type Runtime: Runtime<T, Handle = Self>

Required Methods

fn create_proxy(&self) -> <Self::Runtime as Runtime<T>>::EventLoopProxy

Creates an EventLoopProxy that can be used to dispatch user events to the main event loop.

fn set_activation_policy( &self, activation_policy: ActivationPolicy, ) -> Result<()>

Sets the activation policy for the application.

fn request_exit(&self, code: i32) -> Result<()>

Requests an exit of the event loop.

fn create_window<F: Fn(RawWindow<'_>) + Send + 'static>( &self, pending: PendingWindow<T, Self::Runtime>, before_window_creation: Option<F>, ) -> Result<DetachedWindow<T, Self::Runtime>>

Create a new window.

fn create_webview( &self, window_id: WindowId, pending: PendingWebview<T, Self::Runtime>, ) -> Result<DetachedWebview<T, Self::Runtime>>

Create a new webview.

fn run_on_main_thread<F: FnOnce() + Send + 'static>(&self, f: F) -> Result<()>

Run a task on the main thread.

fn display_handle(&self) -> Result<DisplayHandle<'_>, HandleError>

fn primary_monitor(&self) -> Option<Monitor>

fn monitor_from_point(&self, x: f64, y: f64) -> Option<Monitor>

fn available_monitors(&self) -> Vec<Monitor>

fn cursor_position(&self) -> Result<PhysicalPosition<f64>>

fn set_theme(&self, theme: Option<Theme>)

fn show(&self) -> Result<()>

Shows the application, but does not automatically focus it.

fn hide(&self) -> Result<()>

Hides the application.

Object Safety

This trait is not object safe.