Trait tauri_runtime::WebviewDispatch

pub trait WebviewDispatch<T: UserEvent>:
    Debug
    + Clone
    + Send
    + Sync
    + Sized
    + 'static {
    type Runtime: Runtime<T>;

    // Required methods
    fn run_on_main_thread<F: FnOnce() + Send + 'static>(
        &self,
        f: F,
    ) -> Result<()>;
    fn on_webview_event<F: Fn(&WebviewEvent) + Send + 'static>(
        &self,
        f: F,
    ) -> WebviewEventId;
    fn with_webview<F: FnOnce(Box<dyn Any>) + Send + 'static>(
        &self,
        f: F,
    ) -> Result<()>;
    fn open_devtools(&self);
    fn close_devtools(&self);
    fn is_devtools_open(&self) -> Result<bool>;
    fn url(&self) -> Result<String>;
    fn bounds(&self) -> Result<Rect>;
    fn position(&self) -> Result<PhysicalPosition<i32>>;
    fn size(&self) -> Result<PhysicalSize<u32>>;
    fn navigate(&self, url: Url) -> Result<()>;
    fn print(&self) -> Result<()>;
    fn close(&self) -> Result<()>;
    fn set_bounds(&self, bounds: Rect) -> Result<()>;
    fn set_size(&self, size: Size) -> Result<()>;
    fn set_position(&self, position: Position) -> Result<()>;
    fn set_focus(&self) -> Result<()>;
    fn hide(&self) -> Result<()>;
    fn show(&self) -> Result<()>;
    fn eval_script<S: Into<String>>(&self, script: S) -> Result<()>;
    fn reparent(&self, window_id: WindowId) -> Result<()>;
    fn set_auto_resize(&self, auto_resize: bool) -> Result<()>;
    fn set_zoom(&self, scale_factor: f64) -> Result<()>;
    fn clear_all_browsing_data(&self) -> Result<()>;

}

Webview dispatcher. A thread-safe handle to the webview APIs.

Required Associated Types

type Runtime: Runtime<T>

The runtime this WebviewDispatch runs under.

Required Methods

fn run_on_main_thread<F: FnOnce() + Send + 'static>(&self, f: F) -> Result<()>

Run a task on the main thread.

fn on_webview_event<F: Fn(&WebviewEvent) + Send + 'static>( &self, f: F, ) -> WebviewEventId

Registers a webview event handler.

fn with_webview<F: FnOnce(Box<dyn Any>) + Send + 'static>( &self, f: F, ) -> Result<()>

Runs a closure with the platform webview object as argument.

fn open_devtools(&self)

Open the web inspector which is usually called devtools.

fn close_devtools(&self)

Close the web inspector which is usually called devtools.

fn is_devtools_open(&self) -> Result<bool>

Gets the devtools window’s current open state.

fn url(&self) -> Result<String>

Returns the webview’s current URL.

fn bounds(&self) -> Result<Rect>

Returns the webview’s bounds.

fn position(&self) -> Result<PhysicalPosition<i32>>

Returns the position of the top-left hand corner of the webviews’s client area relative to the top-left hand corner of the window.

fn size(&self) -> Result<PhysicalSize<u32>>

Returns the physical size of the webviews’s client area.

fn navigate(&self, url: Url) -> Result<()>

Navigate to the given URL.

fn print(&self) -> Result<()>

Opens the dialog to prints the contents of the webview.

fn close(&self) -> Result<()>

Closes the webview.

fn set_bounds(&self, bounds: Rect) -> Result<()>

Sets the webview’s bounds.

fn set_size(&self, size: Size) -> Result<()>

Resizes the webview.

fn set_position(&self, position: Position) -> Result<()>

Updates the webview position.

fn set_focus(&self) -> Result<()>

Bring the window to front and focus the webview.

fn hide(&self) -> Result<()>

Hide the webview

fn show(&self) -> Result<()>

Show the webview

fn eval_script<S: Into<String>>(&self, script: S) -> Result<()>

Executes javascript on the window this WindowDispatch represents.

fn reparent(&self, window_id: WindowId) -> Result<()>

Moves the webview to the given window.

fn set_auto_resize(&self, auto_resize: bool) -> Result<()>

Sets whether the webview should automatically grow and shrink its size and position when the parent window resizes.

fn set_zoom(&self, scale_factor: f64) -> Result<()>

Set the webview zoom level

fn clear_all_browsing_data(&self) -> Result<()>

Clear all browsing data for this webview.

Object Safety

This trait is not object safe.