Trait tauri_runtime::webview::WindowBuilder

pub trait WindowBuilder: WindowBuilderBase {

    fn new() -> Self;
    fn with_config(config: WindowConfig) -> Self;
    fn menu(self, menu: Menu) -> Self;
    fn center(self) -> Self;
    fn position(self, x: f64, y: f64) -> Self;
    fn inner_size(self, width: f64, height: f64) -> Self;
    fn min_inner_size(self, min_width: f64, min_height: f64) -> Self;
    fn max_inner_size(self, max_width: f64, max_height: f64) -> Self;
    fn resizable(self, resizable: bool) -> Self;
    fn title<S: Into<String>>(self, title: S) -> Self;
    fn fullscreen(self, fullscreen: bool) -> Self;
    fn focus(self) -> Self;
    fn maximized(self, maximized: bool) -> Self;
    fn visible(self, visible: bool) -> Self;
    fn transparent(self, transparent: bool) -> Self;
    fn decorations(self, decorations: bool) -> Self;
    fn always_on_top(self, always_on_top: bool) -> Self;
    fn icon(self, icon: Icon) -> Result<Self>;
    fn skip_taskbar(self, skip: bool) -> Self;
    fn theme(self, theme: Option<Theme>) -> Self;
    fn has_icon(&self) -> bool;
    fn get_menu(&self) -> Option<&Menu>;

}

A builder for all attributes related to a single webview.

This trait is only meant to be implemented by a custom Runtime and not by applications.

Required Methods

fn new() -> Self

Initializes a new window attributes builder.

fn with_config(config: WindowConfig) -> Self

Initializes a new webview builder from a WindowConfig

fn menu(self, menu: Menu) -> Self

Sets the menu for the window.

fn center(self) -> Self

Show window in the center of the screen.

fn position(self, x: f64, y: f64) -> Self

The initial position of the window’s.

fn inner_size(self, width: f64, height: f64) -> Self

Window size.

fn min_inner_size(self, min_width: f64, min_height: f64) -> Self

Window min inner size.

fn max_inner_size(self, max_width: f64, max_height: f64) -> Self

Window max inner size.

fn resizable(self, resizable: bool) -> Self

Whether the window is resizable or not.

fn title<S: Into<String>>(self, title: S) -> Self

The title of the window in the title bar.

fn fullscreen(self, fullscreen: bool) -> Self

Whether to start the window in fullscreen or not.

fn focus(self) -> Self

Whether the window will be initially hidden or focused.

fn maximized(self, maximized: bool) -> Self

Whether the window should be maximized upon creation.

fn visible(self, visible: bool) -> Self

Whether the window should be immediately visible upon creation.

fn transparent(self, transparent: bool) -> Self

Whether the the window should be transparent. If this is true, writing colors with alpha values different than 1.0 will produce a transparent window.

fn decorations(self, decorations: bool) -> Self

Whether the window should have borders and bars.

fn always_on_top(self, always_on_top: bool) -> Self

Whether the window should always be on top of other windows.

fn icon(self, icon: Icon) -> Result<Self>

Sets the window icon.

fn skip_taskbar(self, skip: bool) -> Self

Sets whether or not the window icon should be added to the taskbar.

fn theme(self, theme: Option<Theme>) -> Self

Forces a theme or uses the system settings if None was provided.

fn has_icon(&self) -> bool

Whether the icon was set or not.

fn get_menu(&self) -> Option<&Menu>

Gets the window menu.