Module tauri_runtime::window

A layer between raw Runtime webview windows and Tauri.

Modules

dpi

UI scaling utilities.

Structs

DetachedWindow

A webview window that is not yet managed by Tauri.

JsEventListenerKey

Key for a JS event listener.

MenuEvent

A menu event.

PendingWindow

A webview window that has yet to be built.

Enums

CursorIcon

Describes the appearance of the mouse cursor.

FileDropEvent

The file drop event payload.

WindowEvent

An event from a window.

Functions

assert_label_is_valid

is_label_valid