Enum tauri_runtime::window::CursorIcon

#[non_exhaustive]
pub enum CursorIcon {

    Default,
    Crosshair,
    Hand,
    Arrow,
    Move,
    Text,
    Wait,
    Help,
    Progress,
    NotAllowed,
    ContextMenu,
    Cell,
    VerticalText,
    Alias,
    Copy,
    NoDrop,
    Grab,
    Grabbing,
    AllScroll,
    ZoomIn,
    ZoomOut,
    EResize,
    NResize,
    NeResize,
    NwResize,
    SResize,
    SeResize,
    SwResize,
    WResize,
    EwResize,
    NsResize,
    NeswResize,
    NwseResize,
    ColResize,
    RowResize,

}

Describes the appearance of the mouse cursor.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Default

The platform-dependent default cursor.

Crosshair

A simple crosshair.

Hand

A hand (often used to indicate links in web browsers).

Arrow

Self explanatory.

Move

Indicates something is to be moved.

Text

Indicates text that may be selected or edited.

Wait

Program busy indicator.

Help

Help indicator (often rendered as a “?”)

Progress

Progress indicator. Shows that processing is being done. But in contrast with “Wait” the user may still interact with the program. Often rendered as a spinning beach ball, or an arrow with a watch or hourglass.

NotAllowed

Cursor showing that something cannot be done.

ContextMenu

Cell

VerticalText

Alias

Copy

NoDrop

Grab

Indicates something can be grabbed.

Grabbing

Indicates something is grabbed.

AllScroll

ZoomIn

ZoomOut

EResize

Indicate that some edge is to be moved. For example, the ‘SeResize’ cursor is used when the movement starts from the south-east corner of the box.

NResize

NeResize

NwResize

SResize

SeResize

SwResize

WResize

EwResize

NsResize

NeswResize

NwseResize

ColResize

RowResize