Struct tauri_runtime::window::PendingWindow

pub struct PendingWindow<T: UserEvent, R: Runtime<T>> {
    pub label: String,
    pub window_builder: <R::Dispatcher as Dispatch<T>>::WindowBuilder,
    pub webview_attributes: WebviewAttributes,
    pub uri_scheme_protocols: HashMap<String, Box<dyn Fn(&HttpRequest) -> Result<HttpResponse, Box<dyn Error>> + Send + Sync + 'static>>,
    pub ipc_handler: Option<WebviewIpcHandler<T, R>>,
    pub menu_ids: Arc<Mutex<HashMap<MenuHash, MenuId>>>,
    pub js_event_listeners: Arc<Mutex<HashMap<JsEventListenerKey, HashSet<u32>>>>,
    pub navigation_handler: Option<Box<dyn Fn(Url) -> bool + Send>>,
    pub web_resource_request_handler: Option<Box<dyn Fn(&HttpRequest, &mut HttpResponse) + Send + Sync>>,
    pub url: String,
}

A webview window that has yet to be built.

Fields
label: String

The label that the window will be named.

window_builder: <R::Dispatcher as Dispatch<T>>::WindowBuilder

The WindowBuilder that the window will be created with.

webview_attributes: WebviewAttributes

The WebviewAttributes that the webview will be created with.

uri_scheme_protocols: HashMap<String, Box<dyn Fn(&HttpRequest) -> Result<HttpResponse, Box<dyn Error>> + Send + Sync + 'static>>ipc_handler: Option<WebviewIpcHandler<T, R>>

How to handle IPC calls on the webview window.

menu_ids: Arc<Mutex<HashMap<MenuHash, MenuId>>>

Maps runtime id to a string menu id.

js_event_listeners: Arc<Mutex<HashMap<JsEventListenerKey, HashSet<u32>>>>

A HashMap mapping JS event names with associated listener ids.

navigation_handler: Option<Box<dyn Fn(Url) -> bool + Send>>

A handler to decide if incoming url is allowed to navigate.

web_resource_request_handler: Option<Box<dyn Fn(&HttpRequest, &mut HttpResponse) + Send + Sync>>url: String

The resolved URL to load on the webview.

Implementations

impl<T: UserEvent, R: Runtime<T>> PendingWindow<T, R>

pub fn new( window_builder: <R::Dispatcher as Dispatch<T>>::WindowBuilder, webview_attributes: WebviewAttributes, label: impl Into<String> ) -> Result<Self>

Create a new PendingWindow with a label and starting url.

pub fn with_config( window_config: WindowConfig, webview_attributes: WebviewAttributes, label: impl Into<String> ) -> Result<Self>

Create a new PendingWindow from a WindowConfig with a label and starting url.

pub fn set_menu(self, menu: Menu) -> Self

pub fn register_uri_scheme_protocol<N: Into<String>, H: Fn(&HttpRequest) -> Result<HttpResponse, Box<dyn Error>> + Send + Sync + 'static>( &mut self, uri_scheme: N, protocol: H )