Crate tauri_runtime

Internal runtime between Tauri and the underlying webview runtime.

Modules

http

menu

Create window and system tray menus.

monitor

Types useful for interacting with a user’s monitors.

webview

Items specific to the Runtime’s webview.

window

A layer between raw Runtime webview windows and Tauri.

Structs

Icon

Window icon.

RunIteration

Metadata for a runtime event loop iteration on run_iteration.

Enums

ActivationPolicy

Application’s activation policy. Corresponds to NSApplicationActivationPolicy.

Error

ExitRequestedEventAction

Action to take when the event loop is about to exit

RunEvent

Event triggered on the event loop run.

SystemTrayEvent

A system tray event.

UserAttentionType

Type of user attention requested on a window.

Traits

Dispatch

Webview dispatcher. A thread-safe handle to the webview API.

EventLoopProxy

Runtime

The webview runtime interface.

RuntimeHandle

A Send handle to the runtime.

UserEvent

A type that can be used as an user event.

Type Definitions

Result

Result type.

TrayEventHandler

TrayId