Trait tauri_runtime::Runtime

pub trait Runtime<T: UserEvent>: Debug + Sized + 'static {
    type Dispatcher: Dispatch<T, Runtime = Self>;
    type Handle: RuntimeHandle<T, Runtime = Self>;
    type EventLoopProxy: EventLoopProxy<T>;

    fn new() -> Result<Self>;
    fn create_proxy(&self) -> Self::EventLoopProxy;
    fn handle(&self) -> Self::Handle;
    fn create_window(
        &self,
        pending: PendingWindow<T, Self>
    ) -> Result<DetachedWindow<T, Self>>;
    fn set_activation_policy(&mut self, activation_policy: ActivationPolicy);
    fn show(&self);
    fn hide(&self);
    fn run_iteration<F: Fn(RunEvent<T>) + 'static>(
        &mut self,
        callback: F
    ) -> RunIteration;
    fn run<F: FnMut(RunEvent<T>) + 'static>(self, callback: F);
}

The webview runtime interface.

Required Associated Types

type Dispatcher: Dispatch<T, Runtime = Self>

The message dispatcher.

type Handle: RuntimeHandle<T, Runtime = Self>

The runtime handle type.

type EventLoopProxy: EventLoopProxy<T>

The proxy type.

Required Methods

fn new() -> Result<Self>

Creates a new webview runtime. Must be used on the main thread.

fn create_proxy(&self) -> Self::EventLoopProxy

Creates an EventLoopProxy that can be used to dispatch user events to the main event loop.

fn handle(&self) -> Self::Handle

Gets a runtime handle.

fn create_window(
&self,
pending: PendingWindow<T, Self>
) -> Result<DetachedWindow<T, Self>>

Create a new webview window.

fn set_activation_policy(&mut self, activation_policy: ActivationPolicy)

Sets the activation policy for the application. It is set to NSApplicationActivationPolicyRegular by default.

fn show(&self)

Shows the application, but does not automatically focus it.

fn hide(&self)

Hides the application.

fn run_iteration<F: Fn(RunEvent<T>) + 'static>(
&mut self,
callback: F
) -> RunIteration

Runs the one step of the webview runtime event loop and returns control flow to the caller.

fn run<F: FnMut(RunEvent<T>) + 'static>(self, callback: F)

Run the webview runtime.