Trait tauri_runtime::Runtime

pub trait Runtime<T: UserEvent>:
    Debug
    + Sized
    + 'static {
    type WindowDispatcher: WindowDispatch<T, Runtime = Self>;
    type WebviewDispatcher: WebviewDispatch<T, Runtime = Self>;
    type Handle: RuntimeHandle<T, Runtime = Self>;
    type EventLoopProxy: EventLoopProxy<T>;

    // Required methods
    fn new(args: RuntimeInitArgs) -> Result<Self>;
    fn create_proxy(&self) -> Self::EventLoopProxy;
    fn handle(&self) -> Self::Handle;
    fn create_window<F: Fn(RawWindow<'_>) + Send + 'static>(
        &self,
        pending: PendingWindow<T, Self>,
        after_window_creation: Option<F>,
    ) -> Result<DetachedWindow<T, Self>>;
    fn create_webview(
        &self,
        window_id: WindowId,
        pending: PendingWebview<T, Self>,
    ) -> Result<DetachedWebview<T, Self>>;
    fn primary_monitor(&self) -> Option<Monitor>;
    fn monitor_from_point(&self, x: f64, y: f64) -> Option<Monitor>;
    fn available_monitors(&self) -> Vec<Monitor>;
    fn cursor_position(&self) -> Result<PhysicalPosition<f64>>;
    fn set_theme(&self, theme: Option<Theme>);
    fn set_activation_policy(&mut self, activation_policy: ActivationPolicy);
    fn show(&self);
    fn hide(&self);
    fn set_device_event_filter(&mut self, filter: DeviceEventFilter);
    fn run_iteration<F: FnMut(RunEvent<T>) + 'static>(&mut self, callback: F);
    fn run<F: FnMut(RunEvent<T>) + 'static>(self, callback: F);

}

The webview runtime interface.

Required Associated Types

type WindowDispatcher: WindowDispatch<T, Runtime = Self>

The window message dispatcher.

type WebviewDispatcher: WebviewDispatch<T, Runtime = Self>

The webview message dispatcher.

type Handle: RuntimeHandle<T, Runtime = Self>

The runtime handle type.

type EventLoopProxy: EventLoopProxy<T>

The proxy type.

Required Methods

fn new(args: RuntimeInitArgs) -> Result<Self>

Creates a new webview runtime. Must be used on the main thread.

fn create_proxy(&self) -> Self::EventLoopProxy

Creates an EventLoopProxy that can be used to dispatch user events to the main event loop.

fn handle(&self) -> Self::Handle

Gets a runtime handle.

fn create_window<F: Fn(RawWindow<'_>) + Send + 'static>( &self, pending: PendingWindow<T, Self>, after_window_creation: Option<F>, ) -> Result<DetachedWindow<T, Self>>

Create a new window.

fn create_webview( &self, window_id: WindowId, pending: PendingWebview<T, Self>, ) -> Result<DetachedWebview<T, Self>>

Create a new webview.

fn primary_monitor(&self) -> Option<Monitor>

fn monitor_from_point(&self, x: f64, y: f64) -> Option<Monitor>

fn available_monitors(&self) -> Vec<Monitor>

fn cursor_position(&self) -> Result<PhysicalPosition<f64>>

fn set_theme(&self, theme: Option<Theme>)

fn set_activation_policy(&mut self, activation_policy: ActivationPolicy)

Sets the activation policy for the application.

fn show(&self)

Shows the application, but does not automatically focus it.

fn hide(&self)

Hides the application.

fn set_device_event_filter(&mut self, filter: DeviceEventFilter)

Change the device event filter mode.

Since the DeviceEvent capture can lead to high CPU usage for unfocused windows, tao will ignore them by default for unfocused windows on Windows. This method allows changing the filter to explicitly capture them again.

Platform-specific

** Linux / macOS / iOS / Android**: Unsupported.

fn run_iteration<F: FnMut(RunEvent<T>) + 'static>(&mut self, callback: F)

Runs an iteration of the runtime event loop and returns control flow to the caller.

fn run<F: FnMut(RunEvent<T>) + 'static>(self, callback: F)

Run the webview runtime.

Object Safety

This trait is not object safe.