Trait tauri_runtime::RuntimeHandle

pub trait RuntimeHandle<T: UserEvent>: Debug + Clone + Send + Sync + Sized + 'static {
    type Runtime: Runtime<T, Handle = Self>;

    // Required methods
    fn create_proxy(&self) -> <Self::Runtime as Runtime<T>>::EventLoopProxy;
    fn create_window(
        &self,
        pending: PendingWindow<T, Self::Runtime>
    ) -> Result<DetachedWindow<T, Self::Runtime>>;
    fn run_on_main_thread<F: FnOnce() + Send + 'static>(
        &self,
        f: F
    ) -> Result<()>;
    fn raw_display_handle(&self) -> RawDisplayHandle;
    fn show(&self) -> Result<()>;
    fn hide(&self) -> Result<()>;
}

A Send handle to the runtime.

Required Associated Types

type Runtime: Runtime<T, Handle = Self>

Required Methods

fn create_proxy(&self) -> <Self::Runtime as Runtime<T>>::EventLoopProxy

Creates an EventLoopProxy that can be used to dispatch user events to the main event loop.

fn create_window( &self, pending: PendingWindow<T, Self::Runtime> ) -> Result<DetachedWindow<T, Self::Runtime>>

Create a new webview window.

fn run_on_main_thread<F: FnOnce() + Send + 'static>(&self, f: F) -> Result<()>

Run a task on the main thread.

fn raw_display_handle(&self) -> RawDisplayHandle

fn show(&self) -> Result<()>

Shows the application, but does not automatically focus it.

fn hide(&self) -> Result<()>

Hides the application.