Module tauri_runtime::window

A layer between raw Runtime windows and Tauri.

Structs

DetachedWindow

A window that is not yet managed by Tauri.

PendingWindow

A window that has yet to be built.

RawWindow

A raw window type that contains fields to access the HWND on Windows, gtk::ApplicationWindow on Linux and NSView on macOS.

WindowId

Identifier of a window.

WindowSizeConstraints

Window size constraints

Enums

CursorIcon

Describes the appearance of the mouse cursor.

DragDropEvent

The drag drop event payload.

WebviewEvent

An event from a window.

WindowEvent

An event from a window.

Traits

WindowBuilder

A builder for all attributes related to a single window.

WindowBuilderBase

Do NOT implement this trait except for use in a custom Runtime

Functions

assert_label_is_valid

is_label_valid